refactor: move RpcError and RpcResult to util
