+++ src/util.rs
@@ -1,4 +1,5 @@
 use std::{
+    result,
     fmt::{self, Display},
     str::FromStr,
 };
@@ -117,6 +118,27 @@ impl<'de> de::Visitor<'de> for HashVisitor {
 impl MaxSize for Hash {
     const POSTCARD_MAX_SIZE: usize = 32;
 }
+
+/// A serializable error type for use in RPC responses.
+#[derive(Serialize, Deserialize, Debug)]
+pub struct RpcError(serde_error::Error);
+
+impl std::error::Error for RpcError {}
+
+impl fmt::Display for RpcError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        fmt::Debug::fmt(self, f)
+    }
+}
+
+impl From<anyhow::Error> for RpcError {
+    fn from(e: anyhow::Error) -> Self {
+        RpcError(serde_error::Error::new(&*e))
+    }
+}
+
+pub type RpcResult<T> = result::Result<T, RpcError>;
+
 
 #[cfg(test)]
 mod tests {
